# docs/contracts/gateway-openapi.yaml
openapi: 3.0.3
info:
  title: WikiAI Gateway API
  version: 0.1.0
  description: Public, admin, health, and external API contract for WikiAI Gateway.
servers:
  - url: http://127.0.0.1:3000
paths:
...
  /api/search:
    post:
      summary: Search allowed MediaWiki content.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SearchRequest"
      responses:
        "200":
          description: Search response with sources and diagnostics.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/chat:
    post:
      summary: Chat over allowed MediaWiki content.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatRequest"
      responses:
        "200":
          description: Chat response or SSE stream depending on request.
        "400":
          $ref: "#/components/responses/BadRequest"
...
  /api/v1/search:
    post:
      summary: Stable external search API.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SearchRequest"
      responses:
        "200":
          description: External search response.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchResponse"
        "409":
          description: Requested retrieval profile is not ready.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/chat:
    post:
      summary: Stable external chat API.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatRequest"
      responses:
        "200":
          description: External chat response.
        "409":
          description: Requested retrieval profile is not ready.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          $ref: "#/components/responses/Forbidden"
...
components:
  securitySchemes:
    mediaWikiSession:
      type: apiKey
      in: cookie
      name: mediawiki_session
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    internalAdminToken:
      type: apiKey
      in: header
      name: x-wikiai-admin-token
  responses:
    BadRequest:
      description: Invalid request.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    Forbidden:
      description: Caller is not allowed to use the endpoint.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    SearchRequest:
      type: object
      required: [query]
      properties:
        query:
          type: string
          minLength: 1
        topK:
          type: integer
          minimum: 1
        retrievalProfileId:
          type: string
          description: Optional admin-defined retrieval profile id. Raw BM25/ColBERT/trigram flags are not accepted by the public API.
        context:
          $ref: "#/components/schemas/ExternalContext"
    SearchChunk:
      type: object
      required: [title, text, score]
      properties:
        title:
          type: string
        pageUrl:
          type: string
        text:
          type: string
        score:
          type: number
        trust:
          type: object
          additionalProperties: true
    SearchResponse:
      type: object
      properties:
        results:
          type: array
          items:
            $ref: "#/components/schemas/SearchChunk"
        chunks:
          type: array
          items:
            $ref: "#/components/schemas/SearchChunk"
        diagnostics:
          type: object
          additionalProperties: true
    ChatRequest:
      type: object
      required: [message]
      properties:
        message:
          type: string
          minLength: 1
        conversationId:
          type: string
        stream:
          type: boolean
        topK:
          type: integer
          minimum: 1
        retrievalProfileId:
          type: string
          description: Optional admin-defined retrieval profile id. Raw BM25/ColBERT/trigram flags are not accepted by the public API.
        context:
          $ref: "#/components/schemas/ExternalContext"
    ExternalContext:
      type: object
      description: Optional caller context for embedded integrations. Gateway validates shape but derives auth server-side.
      properties:
        sourceApp:
          type: string
        objectType:
          type: string
        objectId:
          type: string
        title:
          type: string
        url:
          type: string
        tags:
          type: array
          items:
            type: string
        params:
          type: object
          additionalProperties: true
        attributes:
          type: object
          additionalProperties: true
        dynamicBlocks:
          type: array
          items:
            $ref: "#/components/schemas/DynamicBlockContext"
    DynamicBlockContext:
      type: object
      properties:
        sourceApp:
          type: string
          enum: [cmdbdynamicpages]
        templateCode:
          type: string
        status:
          type: string
          enum: [snapshot_hit, snapshot_miss, auth_runtime_hit, auth_required, permission_denied, runtime_error, unresolved_params]
        title:
          type: string
        url:
          type: string
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string